UI Development
Brief
This section will entail Completion of the UI elements and menus. There will be three headings, one will be the wireframes that where the result of me using unity to make the menus. The second will be the testing of these UI elements, and the final third will be the look and color of all the screens.