Finished adding all 10 levels
New Levels and Fixes
Brief
The first step will be to make at least 10 levels that can be in published version. These need to develop inside separate json files and be correctly format, they will number from map one to ten.
Each one should be play tested to see if it can be completed but this will be done as I record as bugs occur.
Any updates should also be listed under the table and can be from anywhere during implementation of the new levels.