Tile Theme
This section will detail the tile theme and general UI improvements. The new look will have distinct colours, so it doesn’t confuse the user. Then the colours will then have a bloom effect to give it a more techno retro look. The entire game will have a black background a neon glow.
The final improvement is some general UI improvements to help the user figure out the game and the correct way to play it.