State Machines
This will be in place of the Boolean variables that control functions to change the active visibility of the GameObjects. All these systems will be replaced with enumeration variables, that when changed will set the active visibility. This will be done on the condition that it is not already active or unactive to avoid duplicate execution during runtime.
There is an order the states variables and their update if statements.
1.	Game States
2.	Menu States
3.	Mouse States
4.	Run time States