Tutorial
The start button has been replaced with tutorial this button will go into a separate static map that has 2 new arrays. The first will dictate the dialogue shown and the second will be whether the mirror button or the count tile is flashing. With these 2 combine I will be able to construct a tutorial for just one level to help explain the game.