Tutorial 2: Text Complete
This iteration will complete off a Tutorial ready for a test release on the Play Store. The UI will have two elements added to it with the column and row restrictions since they are always different. These will also be included in the flashing elements for the Tutorial and many bugs have been fixed in this iteration.